requirements.txt
pandas==1.0.5
miriam_translator
